allow token:x-oauth-basic basic auth for lfs
We get the user's access token as user and "x-oauth-basic" as password. The "main" repository module already accepts this type of token so lfs should too.

drone.io uses this so lfs access to private repos is now granted and avoids the regression from #5657 for my drone scenario (since drone with the default git plugin requests are already authenticated).